% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{abund_to_occ}
\alias{abund_to_occ}
\title{From abundances matrix to occupancies matrix}
\usage{
abund_to_occ(abundances)
}
\arguments{
\item{abundances}{Abundances matrix}
}
\value{
3D occupancies matrix (species x patch x time)
}
\description{
Turns abundances matrix to occupancies matrix  (species x patch x time)
}
\examples{
# occupancies = abund_to_occ(abundances)

}
