% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_setup.R
\name{env_generate_wrp}
\alias{env_generate_wrp}
\title{Generate Environment (wrapper)}
\usage{
env_generate_wrp(
  landscape,
  env.df,
  env1Scale = 1,
  temporal_autocorr = TRUE,
  timesteps = 1000,
  plot = TRUE
)
}
\arguments{
\item{landscape}{landscape generated by landscape_generate()}

\item{env.df}{optional dataframe with environmental conditions with columns: env1, patch, time}

\item{env1Scale}{scale of environmental autocorrelation between 0 and 1000}

\item{temporal_autocorr}{(logical) adding temporal (autocorrelated) heterogeneity}

\item{timesteps}{number of timesteps to simulate}

\item{plot}{option to show plot of environmental variation}
}
\value{
dataframe with x and y coordinates, time, and environmental conditions
}
\description{
Generates density independent environmental conditions for metacommunity simulations.
Function is here updated - allow to remove temporal environmental heterogeneity
}
\examples{
# env_generate(landscape_generate())


}
\author{
Patrick L. Thompson (updated M.D.)
}
