% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{get_envt}
\alias{get_envt}
\title{Get environmental values from sim. output}
\usage{
get_envt(sim_output)
}
\arguments{
\item{sim_output}{Output from simulate_MC()}
}
\value{
Environmental variable dataframe (patch x envt)
}
\description{
Shortcut to get environmental value, consider there is no temporal variability for envt.
}
\examples{
# data_envt = get_envt(sim_output)

}
