% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{get_geoposition}
\alias{get_geoposition}
\title{Get spatial positions from sim. output}
\usage{
get_geoposition(sim_output)
}
\arguments{
\item{sim_output}{Output from simulate_MC()}
}
\value{
Position dataframe (patch x 2 - (x,y)-coord.)
}
\description{
Shortcut to get spatial positions
}
\examples{
# data_geo = get_geoposition(sim_output)

}
