% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{plots_envt}
\alias{plots_envt}
\title{Plots environmental conditions}
\usage{
plots_envt(sim_output)
}
\arguments{
\item{sim_output}{Output from simulate_MC()}
}
\value{
ggplot plot
}
\description{
Plots environmental conditions in space
}
\examples{
# out = simulate_MC(patches = 100, species = 3)
# plot_envt(out)

}
