% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{plots_occupancies}
\alias{plots_occupancies}
\title{Plots occupancies along time}
\usage{
plots_occupancies(occupancies)
}
\arguments{
\item{occupancies}{Occupancies matrix}
}
\value{
none
}
\description{
Simply plots proportion of occupied patch by species as function of time
}
\examples{
# plots_occupancies(occupancies)

}
