% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{sim_to_matrix}
\alias{sim_to_matrix}
\title{Reformating simulate_MC() output to abundances matrix}
\usage{
sim_to_matrix(sim_output)
}
\arguments{
\item{sim_output}{Output from simulate_MC()}
}
\value{
3D abundances matrix (species x patch x time)
}
\description{
Reformats simulate_MC() output to 3D matrix (species x patch x time)
}
\examples{
# output = simulate_MC(5,5)
# abundances = sim_to_matrix(output)

}
